+++ CHANGES.md
@@ -6,6 +6,7 @@ Expyriment Release Notes
 upcoming (Version 0.8.0)
 ------------------------
 New Features:
+- Antialiasing option for stimuli.Ellipse() and stimuli.Circle()
 - new class: misc.HSVColour. Class to handle HSV colours [hue, saturation, value] 
 - new function: quiting wait or event loops by callback_function, if this function
   returs an instance of the new class control.CallbackQuitEvent
@@ -40,25 +41,30 @@ New Features:
   max_width parameter has been defined
 - new feature: too long words in text boxes will be trimmed automatically, 
   this function can be switch off
-- bug fix anti_aliasing shapes
+- control.defaults.blocking_mode for setting the blocking_mode
+- new feature: improved functionality of ranomize.shuffle_list
 
 Changed:
+- ATTENTION: Open_GL is now also used in window mode and will not be switch
+  off automatically, if Expyriment is not running in fullscreen modus.
 - ParallelPort: Polling will now return 4 bits, including also the status of
   the Busy status line
   (ATTENTION: Byte order in the output of other status lines has changed!)
 - changes at Simon example
 - the property Shape.size has been renamed to shape.shape_size
+- io.screen has new parameter to change the type of sync_screen blocking
 
 Fixed:
-- bug in io.TouchScreenButton crash if duration expired 
+- bug in io.TouchScreenButton crash if duration expired
 - bug in command line interface: order of argument is now irrelevant
 - keypad bug in TextInput
 - unicode bug in TextMenu
+- bug in anti_aliasing shapes
 
 Version 0.7.0 (2 Mar 2014)
 --------------------------
 New Features:
-- new feature in testsuite: Font viewer 
+- new feature in testsuite: Font viewer
 - new extra stimulus: stimuli.extras.RandomDotKinematogram
 - new timer and experiment clock to ensure monotonic timing
 - Clock: new method (static) monotonic_time (this time should be always used)